[project @ Arch-1:robey@lag.net--2003-public%secsh--dev--1.0--patch-26] Transport constructor can take hostname or address tuple part of an ongoing attempt to make "simple" versions of some of the API calls, so you can do common-case operations with just a few calls:
Transport's constructor will now let you pass in a string or tuple instead
of a socket-like object.  if you pass in a string, it assumes the string is
a hostname (with optional ":port" segment) and turns that into an address
tuple.  if you pass in a tuple, it assumes it's an address tuple.  in both
cases, it then creates a socket, connects to the given address, and then
continues as if that was the socket passed in.

the idea being that you can call Transport('example.com') and it will do
the right thing.