[project @ Arch-1:robey@lag.net--2003-public%secsh--dev--1.0--patch-27] add BufferedFile abstraction SFTP client mode is mostly functional.  there are probably still some bugs but most of the operations on "file" objects have survived my simple tests.
BufferedFile wraps a simpler stream in something that looks like a python
file (and can even handle seeking if the stream underneath supports it).
it's meant to be subclassed.  most of it is ripped out of what used to be
ChannelFile so i can reuse it for sftp -- ChannelFile is now tiny.

SFTP and Message are now exported.

fixed util.format_binary_line to not quote spaces.