[project @ Arch-1:robey@lag.net--2003-public%secsh--dev--1.0--patch-60] limit read/write requests to 32KB, advertise 32KB max packet size one of the unit tests was failing because the openssh sftp server was dropping the connection without any error.  turns out they have a maximum allowed write size (possibly around 64KB).  the sftp rfcs have a small hint that some servers may drop read/write requests of greater than 32KB.
so, all reads are limited to 32KB, and all writes > 32KB are now chopped up
and sent in 32KB chunks.  this seems to keep openssh happy.

also, we now advertise 32KB max packet size instead of 8KB (the speed
improves a lot), and log when we read/write a packet.  and sftp files are
flushed on seek.