Remove comparison between int and str
The code had been doing 'n < self.__in_buffer' when it
wanted to be doing 'n < len(self.__in_buffer)'

In Python 2.x, this comparison (int < str) is always True.
I found this while porting to Python 3 where it raises
an error.

The code has been working without complaints because always
taking the true branch of this conditional is actually fine.
We don't need the false branch, so drop the check entirely.