Give up on AddressFamily-specific lookups.
Resort to calling getaddrinfo without specifying address family, and
instead only set the FQDN if a result matches the configured
AddressFamily. So, it could still be slow, but this, combined with
BindAddress, should allow the selection of a specific source host to use
in IdentityFile, etc.

Would really like to know how the Travis worker VMs are configured....