Correct undefined name error in LazyFqdn.__str__
If AddressFamily is set to anything but "any" in your ssh config, this (presumably little-used; it's a fluke of history that we have that enabled in our environment) code path runs, and the undefined "host" variable gets referenced. I'm guessing, at some point in the past, host used to be set to socket.gethostname(); here's a quick patch. :)