Allow two-factor auth and unlocking of SSH keys at the same time
As of Paramiko 1.11.0 the user can either authenticate against a server with
two-factor auth _or_ unlock an encrypted key with a provided password. Also,
the user can unlock a private SSH key and perform two-factor auth if the
passphrase of the key and the server logon password are identical.

A more generic approach is supporting a passphrase=... and a password=...
parameter in paramiko.client.SSHClient().connect(). The passphrase (if
given) is used for unlocking a key, the password is used for everything
else (i.e. also for unlocking the key in case passphrase is omitted).

If no passphrase is given, the password is used for unlocking the private
key (to sustain the behvaiour of earlier paramiko versions).