Take known_hosts into account when sending the preferrerd key order.
A common problem when using parameter on systems like Ubuntu or Debian
is that the default host key request by the regular SSH client is
ecdsa-sha2-nistp256, while Paramiko will ask for ssh-rsa first. This will
cause a miss upon the first key lookup and will trigger an "host is not
in known_hosts" warning, even though it's present.

This patch will check if the host is known first and alter the list of
preferred host key types based upon that information before connecting.