Tweak BER encoding to be DER for SSH keys
Do not add sign padding to a data value's content length in long
definite form so it can be decoded as DER.

Paramiko, like OpenSSH, consumes private SSH keys in BER, but unlike
OpenSSH, and against Postel's Law (ie decode liberally and encode
conservatively), it produces keys in BER, not DER (a stricter subset
of BER).  This breaks tools that consume private SSH keys and only
implement DER decoding, tools that are compatible with OpenSSH output.

The difference between Paramiko's BER and OpenSSH's DER (as
implemented by OpenSSL), for the purposes of SSH RSA and DSA private
keys, is the number of subsequent length octets in the long definite
form.  Paramiko includes sign padding which sometimes introduces an
extra, unnecessary octet.  This octet is unnecessary because content
length is not encoded as two's complement and therefor doesn't require
a leading sign bit, but as it is always non-negative and only ever
produces a leading octet of zeros, ie zeros in the most significant
bits, it doesn't alter the content length value and therefor
incidentally legal in BER (see ITU-T X.690 (11/2008), section
8.1.3.5).  While being extra and unnecessary for BER, the octet
prevents private SSH keys output by Paramiko from being decoded as DER
because for DER "length encoding shall be... encoded in the minimum
number of octets" (see ITU-T X.690 (11/2008), section 10.1).