Update sftp_client.py and sftp_file.py so that STAT on a file is only called once and the value passed to any required subroutines.
This is to allow support for the IBM Sterling Integrator SFTP Server when files are defined with an extractability count of 1.

Files with an extractability count of 1 will become unavailable once the GET request has started and any subsequent STAT commands to obtain the file size will fail with a file not found error.