Modified the order of the _preferred_keys to indicate preference to the ecdsa keys
Resent versions of OpenSSH prefers to use ecdsa
keys for the host key verification, and therefore,
ecdsa keys tend to end up in the known_hosts file.
Paramiko on the other hand tend to end up using
rsa keys. This often causes Paramiko to fail the
host key verification despite the fact that the host
is known to OpenSSH. This simple change reduces the
risk of this happening.