Invent the parameter 'gss_trust_dns' for Kerberos support
In response to Paramiko issue #915 the parameter 'gss_trust_dns' was
added for Kerberos support.  Set by default to 'True' the parameter
indicates whether or not the DNS is trusted to securely canonicalize the
hostname of the target host. If set to 'False' the hostname  entered
will be passed to GSSAPI.
This option behaves like GSSAPITrustDNS from OpenSSH.