remove ResourceManager to fix leak of Transport/SSHClient
The back-reference from Transport to SSHClient was added because
the SSHClient had a destructor that would close the Transport,
and some users did not want the Transport closed when the SSHClient
was garbage collected.

The SSHClient destructor was not a normal destructor, it was
implemented with the ResourceManager singleton. Together with
the back-reference, this prevented the GC cycle detector from
freeing the SSHClient and Transport.