Fix zombie process accumulation in ProxyCommand
After a proxy command exits, a defunct process will stick around
until the calling process exits.  This can accumulate to hundreds
or thousands of processes if you are polling a remote server.

The main problem is that wait() wasn't being called, however some
extra logic was added to send a KILL if for some reason TERM doesn't
end up doing it.