Fix failure in listdir when server uses a locale
Fixes #985

SFTPAttributes uses the locale-aware %b directive for the
abbreviated month name with time.strftime, but was not
decoding the result on Python 2.7.

Add a strftime alias in py3compat that will always return
unicode, and resolve the mixing of bytes and text in
SFTPAttributes methods.

Add a test at the listdir level, and a more specific test
for the SFTPAttributes asbytes method.