server: Support pre-authentication banners
The ssh protocol allows for the server to send a pre-authentication
banner. It may be sent any time between the start of authentication and
successful authentication. This commit allow ServerInterface subclasses
to define messages which we'll send right right at the start of
authentication before we send the supported authentication methods.