AuthHandler: fix the server-mode "gssapi-with-mic" logic
A paramiko server is now able to handle a restart of the user
authentication during the GSS-API token exchange. This may occur, if
the client detects a local GSSAPI problem (e.g. a missing kerberos
ticket) and continues with another authentication method.

The added test case test_2_auth_trickledown still fails, because the
paramiko client contains a bug too.