Detect what authentication types are allowed
Some systems are configured to close the connection after one
authentication attempt.  If this is the case and we are provided with
both a pulic key and a password then the password will never be tried.

To fix that, we need to detect the allowed authentication methods and
only try public key or password if those are allowed.