config: Add support for %% tokens escapes.
This add support for %% tokens.  The parser had to be redone since it
would fail with e.g. '%%h' since it rely on a series of str.replace
and would end up interpreting the '%h' token anyway.

Add a new test case for it.

The %% token is necessary for e.g. setting Hostname to an IPv6 link-local
address.