Add safe binding of sha* functions for Kex* classes
    Depending of implementation of hashlib module, sha* functions from
hashlib module can be represented by different objects (to be precise,
with decorated or not decorated functions from _hashlib module).
    In our case, these functions added to classes as class attributes.
It is not safe to call them using self without wrapping with
staticmethod.

Example:
    When type(sha256) is "<class 'builtin_function_or_method'>"
original code will work without an issues, but when type(sha256) is
"<class 'function'>" the code will fail with an exception.

    Test for the case also added.

resolves paramiko/paramiko#1182