Check authenticated status in wait_for_response before raising
change the logic used in auth_handler.wait_for_response exception
raising when transport is inactive

it's possible to enter the auth_handler wait_for response failure
handler during cleanup, if the authentication has completed without
error just as the transport.active flag gets switched (e.g. in
response to a disconnect message. In this state, we have
authenticated: true , but transport.active: false, which causes an
inappropriate AuthenticationException to be raised. The patch just
adds an additional check of the authenticated flag before raising this
error.