BufferedFile: save temp buffer on readline() timeout
Fixes paramiko/paramiko#1196

self._buffer is copied by ref to the line variable, but since bytes
type is immutable, adding data to the line further on, actually makes
it to diverge from self._buffer. Therefore if timeout occurs, the data
that was read during the readline() execution so far, gets lost.

The idea is to save temporary buffer back to the original if we hit timeout.