OrPipe: add proper locking
In the context of fileno(), a Channel's .in_buffer and .in_stderr_buffer
operate on a pair of of OrPipes obtained from make_or_pipe(). These two
cooperate and share the same underlying PosixPipe resp. WindowsPipe.

Due to the lack of any synchronization in the OrPipe's .set() method, it
is possible (and has actually been observed) to have both OrPipes'
._set = True while the underlying pipe remains unset.

Fix that by adding proper locking to the OrPipe implementation.