demos: Use sendall() instead of send() for port forwarding
to avoid losing bytes when send buffer/window fills up

Fixes paramiko/paramiko#511

This fix is not ideal. Since sendall() blocks until all data is sent,
there may be a risk of a deadlock if there is data congestion in both
directions simultaneously. However, the likelihood is significantly
lower than the likelihood of hitting the current bug where bytes are
lost.