Fixes #1255: Add os.path-like stat-based methods to SFTPClient
Add methods to `SFTPClient` that resemble equivalent methods from
`os.path`:
 * `exists`
 * `lexists`
 * `isfile`
 * `islink`
 * `isdir`

All of these methods are based on the results of calling `stat` on a
path, except `lexists`, which is based on `lstat`.

These methods reduce the need for a user of the paramiko library
to import the Python `stat` module and inspect the return value of
an `SFTPClient.stat` call.