handling of terminated proxy commands
The existing code will not robustly detect if the proxy command has terminated;

In 'recv', since we are usually just polling and not performing any operation on the pipe, no operation on the pipe is performed for it to detect/raise an error condition.  'send' is OK, this should raise 'broken pipe', but I believe we should poll the subprocess to check it's still alive within the 'recv' loop.

Also, the 'closed' property currently checks returncode, which would never be set without polling the subprocess anyway.