Make Transport.connect use Transport.auth_none
Makes Transport.connect try to authenticate via method-none first.

This makes the method work with SSH servers that accept method-none as
the only working authentication method.

In order to be backwards-compatible, BadAuthenticationType exceptions
are shielded against and are discarded silently. As a bonus, this makes
the method log the possible authentication methods on a DEBUG level.