clean up SSHException subclass args and string methods
Setting "self.args" is equivalent to calling the parent class
__init__() method with those args, and also overrides the only
thing that parent class __init__() method did. It is confused.

What is actually desired is to override the __str__() method with
a custom formatted exception message. The default for Exception
with zero args is blank, with one arg is the str of that arg,
with multiple args is just the repr of the args tuple.
(That never includes the class name. So, we want custom str.)