issue #1141: avoid quadratic buffer management on Python 3.x
When size is negative or None, it is possible for the result of the read
to grow without bound. Since result is managed through repeat string
concatenation, runtime is quadratic in the number of successful calls to
_read().

In Python 2 this had reasonable runtime because Python/ceval.c contained
a special case for appending to a string that could be safely mutated,
as the only reference to it existed on the stack.
("Python/ceval.c::string_concatenate()"), however during the great
Python 3.x transition this special case was rewritten to only support
Unicode, causing a huge amount of heap churn and memory copies on 3.x.