BufferedFile: avoid quadratic buffer append with Python 3.x
Since result is managed through repeat string
concatenation, runtime is quadratic in the number of successful
calls to read().

In Python 2 this had reasonable runtime because Python/ceval.c contained
a special case for appending to a string that could be safely mutated,
as the only reference to it existed on the stack.
("Python/ceval.c::string_concatenate()"), however during the great
Python 3.x transition this special case was rewritten to only support
Unicode, causing a huge amount of heap churn and memory copies on 3.x.

fixes paramiko/paramiko#1141