transport: gracefully handle EOFError when closing packetizer
Adding exception handling for packetizer race condition when the server ends
the session before the client. This can occur when utilizing fabric2 2.5.0 and
sends a high volume of commands through a Connection() object, then trying to
close the object.

The EOFError is being raised when trying to write out the final bytes to close
the socket, but the socket is already closed. This does not prevent the rest of
the transport object shutdown properly, so a warning log is written out instead
of raising the exception higher.