add filename in ssh key errors
Just getting an error like "Private key file is encrypted" is really
unhelpful: which keyfile? How do I tell which file paramiko tripped
on? I could not figure it out short of running an strace on the
program or injecting a debugger or trace in the source.

I tried to cover all the cases where a private key file is read, which
seem to be where most of the pain is (e.g. in #387).

With this patch, the filename is passed to the read functions, which
is rather unfortunate: it would be better if a file pointer would be
passed into those functions, so that we could just access the `.name`
parameter from there. But it seems that ship has sailed.

Before:

    paramiko.ssh_exception.PasswordRequiredException: Private key file is encrypted

After:

    paramiko.ssh_exception.PasswordRequiredException: Private key file '/home/anarcat/.ssh/id_rsa' is encrypted