Do not send several sshauth-service requests
Some SSH servers do not expect to receive more than one
sshauth-service. The RFCs doesn't seem to say anything definitive about
this, but RFC 4251 says in the introduction:

> The client sends a service request once a secure transport layer
> connection has been established. A second service request is sent
> after user authentication is complete. This allows new protocols to
> be defined and coexist with the protocols listed above.

So, it may be valid to not accept subsequent service request during
authentication.

To avoid sending a sshauth-service request for each auth attempt, we
reuse the AuthService() object for each authentication request to keep
track if we are already in the sshauth service.

This is also more efficient as we do not have an additional round-trip
for each auth tentative.