Add SOCKS5 proxy functionality
This adds SOCKS5 proxy functionality to paramiko as requested and
discussed in #955.

The functionality is nearly identical feature-wise to the one in
OpenSSH, with the notable exception that it only support SOCKS5 and not
SOCKS4. The reasoning behind it being that SOCKS4 was already
superseded by SOCKS5 25 (!) years ago and that SOCKS4 doesn't even
support tunneling requests to hosts using IPv6.

Common implementation details with OpenSSH include:
- supports starting multiple SOCKS5 servers bound to different
  local addresses for a single SSH connection
- exposes the SOCKS5 servers as local IPv4 or IPv6 stream sockets
- only supports a subset of SOCKS5, in particular tunneling of
  UDP-traffic, authentication and incoming connections aren't supported
- supports tunneling connections to IPv4 and IPv6 addresses, as well as
  connections to host names, by resolving them on the SOCKS server side
- 1:1 mapping between SOCKS5 requests and direct-tcpip channels

The implementation works with all Python versions supported by paramiko.
It has so far only been tested with Linux, but should work on other
operating systems as well. The implementation has been used productively
for several weeks so far, using `requests` with its SOCKS5 proxy
functionality as SOCKS5 client, and works without any problems.