Try getting Kerberos working on Circle
Sadly it fails. Unclear whether I'm missing something,
or if this represents a diff between the Ubuntu versions Travis
last used and what Circle's images are on.

I did get the tests themselves truly running, and while I trimmed back
the deps a bit vs Travis, it's only as some of them were
redundant/required by other packages we install here.

Leaving it defined but not actually used in the workflow.