Make changes to allow paramiko to mostly work under FIPS
Right now, because of a couple of invocations of md5() to generate
fingerprints for log messages, paramiko does not work on systems where
FIPS mode is enabled.  This is because hashlib.md5() fails when FIPS
mode is enabled on the kernel.

An attribute has been added to hashlib in python 3.9 to allow
the use of md5() in non-security contexts.  This is because FIPS allows
the use of md5() when secret data is not involved - which is exactly the
case here.

This attribute "usedforsecurity" has been backported to python versions
prior to python 3.9 on some distributions.  We therefore add a check
to confirm whether the attribute is supported, and if so, annotate
accordingly to allow the operation to succeed.

See https://bugs.python.org/issue9216 for more details.