channel: fix race between sending channel EOF and CLOSE messages
While the channel tracks its state for generating messages under
its own lock, it needs to release that lock before calling the
transport method to send a message, which uses the transport's lock.
Messages could be generated under the correct conditions and in the
correct order, but then sent in a different order, if two different
threads are calling channel methods.

(We need to release the channel lock to avoid deadlock during
re-keying according to commit 305f5e09a58d.)

This was made much more likely by the recent addition of
ChannelStdinFile (in paramiko-ng 2.8.0) which calls shutdown_write()
automatically when closed. So, for now, protect only EOF to fix the
immediate regression, and reduce code churn. There is already a mess
of state checking of dubious robustness, e.g. the open_only()
decorator ...

This problem was well described in paramiko/paramiko#1115 and this
solution is heavily inspired by paramiko/paramiko#1137 so:

Co-Authored-By: Ryan <rrasti@google.com>