Fix and improve 'get_hostnames()' for configs with 'Match' directives
SSHConfig.get_hostnames() is unprepared to be called for configs with
'Match' directives, as this function expects every entry in the config
to have a "host" key, which is not the case when the config contains
"Match" directives.

Improve the function by checking for "matches" entries, and
additionnally check if the condition of the 'Match' directive is "host",
in which case add the listed hosts to the set of hosts returned.