Accept capitalized 'Match' conditions
'ssh_config' specifies keywords are case-insensitive, and their value
case-sensitive.

The 'Match' criteria (all, host, canonical, etc.) are listed as
lowercase but are actually also accepted when capitalized by OpenSSH
(maybe because they are labelled criteria "keywords"?).

Be lenient and allow them to be capitalized by making sure to lowercase
them when we read the config. Capitalize one keyword in a test config
for good measure.