Skip tests requiring sha1 signing if the backend doesn't support that
Red Hat Enterprise Linux 9 will have SHA-1 signatures disabled by
default. It is likely that SHA-1 signatures will disappear elsewhere
over time too. This change detects if sha1 signatures are not
supported by the backend and skips tests that rely on that
functionality. This is a workaround for #2004.

It would be good to reduce the reliance of the test suite on sha1
signatures except in the cases where that is explicitly being tested,
and the markers added here give a decent starting point for seeing where
to change things.