Improve performance by reducing expensive bytes conversion
In two core parts of the codebase, complex type-switch based code is used to convert a value to bytes. However, in all cases (except for one, fixed in this PR), the caller is specifying a Message instance. We can make this code much simpler by directly calling the correct method on Message.

In #2110 this is measured to massively speed up large SFTP transfers.