Make includes load in order
This solution is kind of hacky, as it breaks up the context blocks into
parts based on the `Include`, so that they are used in the order of
definition and not at the end of the block.

A proper solution would be to use an ordered collection of options,
during parsing, that ensures ordered lookup during the lookup phase.

OrderedDict was tried as a drop-in, but it is not sufficient enough,
as defined in the include-order-2 test case.