Remove unused deprecated self._config field
The storage of the parsed configs might need to be reviewed,
as the include support introduced complexity.

Due to how includes work, the order of the include is also relevant
during lookup time.

One possible solution would be to flatten the parsed config:
to compile the host/match conditions through includes during parsing.
This way the lookup could be much simpler.