Add skip_host_key_verification option
Adds _skip_host_key_verification attribute to Transport and
provides a set_skip_host_key_verification method to set or clear
it (default: `False`).

Adds skip_host_key_verification parameter to SSHClient.connect
(default: `False`), which is passed directly to the transport.

Skipping host key verification might be useful in very rare cases
where it is not possible to verify the target's key, for example if
it uses a legacy signature algorithm that is no longer supported by
the client, e.g. in #2259.