Support exit-signal messages
See RFC 4254 section 6.10. This is a close sibling of the `exit-status`
message: if the remote process has been terminated due to a signal, the
server may indicate this condition with the `exit-signal` message.

`exit-signal` message handler code uses `Channel.status_event`, as does
the code for the `exit-status` as both messages indicate a termination
of the remote process. It does not make much sense for the server to
send both.

While most servers omit support for `signal_message` and `language_tag`
fields, most library implementations do not, and hence the addition of
these fields here.

All string field arguments / return values are passed as bytestrings,
since the standard does not specify encoding of the `signal_name` nor
the `language_tag`. The signal_message uses `bytes` for consistency.