fix a race in ftp prefetching when using max_concurrent_requests
when `max_concurrent_requests` is set, it's possible to empty the list
of prefetch extents by reading faster than they are enqueued. currently
this makes us mistakenly thing the prefetching is all done. reading
still works, but it starts ignoring the prefetch which, in particular,
means that in the common case of reading an entire file, the prefetch
buffers continue filling but are never emptied causing excessive memory
use.

fix this by explicitly keeping track of when we're done enqueueing.